// frontend/ClientApp/src/modules/template-repository/models/contract-templace.ts
export interface SemanticCondition {
    conditionId: string
    conditionName: string
    schemaVersion: 'v1'
    parameters: SemanticConditionParameter[]
}

export type SemanticParameterType = 'date' | 'string' | 'integer' | 'decimal'

export const SemanticOperate = {
    lessThan: 'lessThan',
    lessThanOrEqual: 'lessThanOrEqual',
    greaterThan: 'greaterThan',
    greaterThanOrEqual: 'greaterThanOrEqual',
    equal: 'equal',
    notEqual: 'notEqual',
} as const

export type SemanticOperateType = (typeof SemanticOperate)[keyof typeof SemanticOperate]

export interface SemanticParameterOperator {
    operate: SemanticOperateType
    /**
     * Target could be other parameters or basic value, 
     * for example: 
     * { "operate": "greaterThan", "value": "100" } 
     * { "operate": "greaterThan", "value": "{{startDate}}" } 
     */
    targets: string[]
}

export interface SemanticConditionParameter {
    parameterName: string
    type: SemanticParameterType
    isRequired: boolean
    operators: SemanticParameterOperator[]
    value: unknown
}